da_agent/requirements.txt
streamlit
pandas
sqlalchemy
plotly
vnstock
llama-index
duckdb-engine
python-dotenv
```